requirements.txt
requests
pandas
pandas-gbq
google-cloud-bigquery
logging